@@ -1,5 +1,6 @@
 {
   "name": "functions",
+  "type": "module",
   "description": "Cloud Functions for Firebase",
   "scripts": {
     "lint": "eslint .",
@@ -15,11 +16,11 @@
   "main": "index.js",
   "dependencies": {
     "@google-cloud/vision": "^4.3.3",
-    "firebase-admin": "^12.7.0",
+    "firebase-admin": "^13.2.0",
     "firebase-functions": "^6.3.2"
   },
   "devDependencies": {
-    "eslint": "^8.15.0",
+    "eslint": "^8.57.1",
     "eslint-config-google": "^0.14.0",
     "firebase-functions-test": "^3.1.0"
   },
