@@ -391,9 +391,44 @@ const gptHomeroomDetailMsg = [
 const gptPersonalOnReportMsg = [
   //1. 대전제
   {
-    role: "system", content: `당신은 과목 세부 능력 특기사항을 객관적 관찰자 시점으로 작성해야하는 교사임.
-    지금부터 생활 기록부의 교과 세부 능력 및 특기사항'을 작성해야함. 문장을 '~음', '~펼침', '~임', '~함', '~됨' 등의 표현으로 끝내야 함.
-    제공되는 활동 문구와 학생의 자기 활동 보고서를 섞어서 개별화 된 문구를 작성해줘. 문맥에 맞게 구체적인 활동 사례와 태도와 사고력에 관한 깊이 있는 평가문을 서술해야함.`
+    role: "system", content: `학생을 객관적으로 관찰한 교사.
+    지금부터 생활 기록부의 교과 세부 능력 및 특기사항'을 작성해야함. 반드시 '~음', '~펼침', '~임', '~함', '~됨'등으로 끝나는 명사형 종결어미를 사용하여 문장을 끝맺어야 함
+    제공되는 활동 문구와 활동 보고서를 적절히 섞어서 새로운 문구를 작성해줘. 문맥에 맞게 구체적인 활동 사례와 태도, 사고력에 관한 깊이 있는 평가문을 작성해야함.
+    학생의 활동 보고서가 혹시 영어라면 한국말로 번역하여 문구를 작성해줘`
+  },
+  //2-8. 예시8
+  {
+    role: "user",
+    content: `활동 문구: 'Lamb to the Slaughter'를 읽고 제목이 주는 의미와 작품의 대표적 기법인 아이러니에 대해 자신의 의견을 작성함.
+    아래는 위의 활동을 한 학생이 쓴 활동 보고서야.
+
+    학생 활동 보고서:나는 '양'이 메리 말로니를 상징한다고 생각해요. 그녀는 매우 친절하고 온순하며 아름다운 주부였어요. 
+    그러나 남편을 죽인 후에는 도살자로 변모했죠. 매우 드라마틱한 일이었어요. 상황적 역설은 형사들이 메리 말로니의 살인을 증거하는 양다리를 먹는 것입니다.
+    
+    위 활동과 보고서를 적절히 섞어서 생기부 문구를 만들어줘, 조건은 '학생'이라는 주어가 없어야 하고 문장을 끝낼 때, 반드시 '~음', '~펼침', '~임', '~함', '~됨'등으로 끝나는 명사형 종결어미를 사용하여 문장을 끝맺어야 함`
+  }, {
+    role: "assistant", content: "Lamb to the Slaughter'를 읽고 제목에서의 '양'은 Mary를 상징한다고 생각함."
+      + " 그녀는 아름다운 가정 주부였으나, 결국 남편을 죽이는 학살자로 변모하였으며 이 부분이 극적이라는 의견을 작성함"
+      + " 작품의 핵심 기법인 상황적 아이러니가 무엇인지 이해하였으며 그 예시로"
+      + " 형사들이 살인 도구인 양 다리를 먹어치웠다는 점을 든 간단한 보고서를 작성함"
+  },
+  //2-7. 예시7
+  {
+    role: "user",
+    content: `활동 문구: 'Lamb to the Slaughter'를 읽고 제목이 주는 의미와 작품의 대표적 기법인 아이러니에 대해 자신의 의견을 작성함.
+    아래는 위의 활동을 한 학생이 쓴 활동 보고서야.
+
+    학생 활동 보고서: 양은 메리를 상징한다고 생각해요. 성경에서 양은 보통 하나님의 순수하고 결백한 아이를 상징합니다. 
+    그러나 소설에서 메리는 살인자가 됩니다. 그래서 양은 도살당하러 가는 것이지만, 양 자체가 도살당하는 존재가 됩니다. 이 소설에는 역설적인 상황들이 있어요. 
+    먼저, 메리가 자신의 남편을 죽입니다. 소설의 시작에서 메리는 남편을 기대하며 기다리지만 나중에는 그녀가 사랑하는 남편을 자신의 손으로 죽입니다. 
+    둘째, 그녀는 양다리를 사용해 그를 죽입니다. 그리고 그녀는 형사들이 증거를 파괴하도록 합니다. 메리가 형사들에게 양을 먹이고 웃는 장면에서 정말 오싹한 느낌이었어요.
+
+    위 활동과 보고서를 적절히 섞어서 생기부 문구를 만들어줘, 조건은 '학생'이라는 주어가 없어야 하고 문장을 끝낼 때, 반드시 '~음', '~펼침', '~임', '~함', '~됨'등으로 끝나는 명사형 종결어미를 사용하여 문장을 끝맺어야 함`
+  }, {
+    role: "assistant", content: "Lamb to the Slaughter'를 읽고 제목에서의 'lamb' 이 순진한 가정주부에서 살인자로 변모한 Mary를 나타낸다고 생각함."
+      + " 또한 작중에서 사용된 아이러니를 학습하고 순수했던 Mary가 살인자로 변한 점과"
+      + " 음식을 살해 도구로 사용한 점 등이 아이러니 하다는 의견을 작성함"
+      + " 마지막에 그녀가 형사들을 속인 부분에서 소름이 돋았다고 느낌을 작성함."
   },
   //2-1. 예시1
   {
@@ -402,9 +437,9 @@ const gptPersonalOnReportMsg = [
     독서를 통해 진로에 대한 흥미와 가능성을 발견하고, 해당 분야에 대한 깊은 이해를 높임. 진로 탐색을 통해 자신의 강점과 흥미를 발견하고,
     독서를 통해 쌓은 지식을 현실적인 진로 선택에 활용함. 함께한 동료들과의 토의를 통해 다양한 진로에 대한 시야를 넓히고, 진로 결정에 도움을 주는 활동을 진행함. 
 
-    아래는 위의 활동을 한 학생이 쓴 진로 보고서야.
+    아래는 위의 활동을 한 학생이 쓴 활동 보고서야.
 
-    활동 보고서: 재정학을 배우면서 얻게 된 의의는 국가 재정 운영이 국가 경제와 사회 전반에 미치는 영향을 깊이 이해하게 되었다는 점이다.
+    학생 활동 보고서: 재정학을 배우면서 얻게 된 의의는 국가 재정 운영이 국가 경제와 사회 전반에 미치는 영향을 깊이 이해하게 되었다는 점이다.
     재정학은 단순히 세금과 예산을 다루는 학문이 아니라, 국가 정책이 경제 및 사회 구조에 어떻게 영향을 미치는지 탐구하는 학문이다. 
     예산 편성이나 공공 재정 정책은 국민들의 생활에 직접적인 영향을 주며, 경제 발전과 사회적 형평성을 달성하는 데 중요한 역할을 한다. 
     재정학을 통해 공공재와 사회적 형평성의 중요성도 깊이 깨닫게 되었다. 공공재는 국민 모두가 함께 누려야 할 가치이며, 
@@ -415,7 +450,7 @@ const gptPersonalOnReportMsg = [
     사회 계층 간의 격차를 줄이는 것은 국가의 지속 가능한 발전에 필수적이다. 이러한 정책을 통해 단순한 경제 성장 이상의 사회적 가치를 실현하는 것이 재정학의 
     중요한 목표임을 이해하게 되었다.
 
-    위 활동과 보고서를 적절히 섞어서 생기부 문구를 만들어줘, 조건은 '학생'이라는 주어가 없어야 하고 '~음', '~펼침', '~임', '~함', '~됨'등으로 끝나는 문장으로 작성해줘`
+    위 활동과 보고서를 적절히 섞어서 생기부 문구를 만들어줘, 조건은 '학생'이라는 주어가 없어야 하고 문장을 끝낼 때, 반드시 '~음', '~펼침', '~임', '~함', '~됨'등으로 끝나는 명사형 종결어미를 사용하여 문장을 끝맺어야 함`
   }, {
     role: "assistant", content: "진로맞춤형 독서 활동을 통해 재정학에 대한 흥미와 가능성을 발견하고, 국가 재정 운영이 경제와 사회 전반에 미치는 영향을 깊이 이해하게 됨."
       + " 재정학은 단순히 세금과 예산을 다루는 학문이 아니라, 공공재와 사회적 형평성의 중요성을 탐구하며, 정부 재정이 경제적 불평등 문제를 해결하고 사회적 안전망을 구축하는"
@@ -429,15 +464,15 @@ const gptPersonalOnReportMsg = [
     독서를 통해 진로에 대한 흥미와 가능성을 발견하고, 해당 분야에 대한 깊은 이해를 높임. 진로 탐색을 통해 자신의 강점과 흥미를 발견하고,
     독서를 통해 쌓은 지식을 현실적인 진로 선택에 활용함. 함께한 동료들과의 토의를 통해 다양한 진로에 대한 시야를 넓히고, 진로 결정에 도움을 주는 활동을 진행함. 
 
-    아래는 위의 활동을 한 학생이 쓴 진로 보고서야.
+    아래는 위의 활동을 한 학생이 쓴 활동 보고서야.
 
-    활동 보고서: 이번 독서 모임에서 나는 보고서 작성과 자료 정리를 맡았다. 팀원들이 모아준 자료를 주제별로 정리하고 취합해 보고서를 작성했다.
+    학생 활동 보고서: 이번 독서 모임에서 나는 보고서 작성과 자료 정리를 맡았다. 팀원들이 모아준 자료를 주제별로 정리하고 취합해 보고서를 작성했다.
     토론이 끝난 후에는 토론 내용을 정리해 기록하였고, 탐구 보고서 등 관련된 문서를 작성하였다. 보고서를 작성한 후에 팀원들의 의견과 피드백을 받아 수정하여
     완성도를 높이는 역할도 맡았다. 우리는 향후 미래의 진로 결정에 도움이 될 만한 책들을 선정하고자 하였다. 책을 읽고 여러 팀원들의 다양한 관련 의견을 들으면서,
     내 희망 진로 분야인 기계 공학자나 로봇 공학자에 대해서 구체적으로 돌아볼 수 있는 계기가 되었다. 책의 내용과 팀원들의 의견을 통해 기계 공학자 또는 로봇 공학자가
     연구할 수 있는 중요한 분야 중 하나인 로봇 의수, 의족, 또는 로봇 건축 등에 대해서 구체적으로 알게 되었고, 나의 꿈의 방향을 명확히 하면서 더 심도 있게 하는 경험이 되었다.
-
-    위 활동과 보고서를 적절히 섞어서 생기부 문구를 만들어줘, 조건은 '학생'이라는 주어가 없어야 하고 '~음', '~펼침', '~임', '~함', '~됨'등으로 끝나는 문장으로 작성해줘`
+    
+    위 활동과 보고서를 적절히 섞어서 생기부 문구를 만들어줘, 조건은 '학생'이라는 주어가 없어야 하고 문장을 끝낼 때, 반드시 '~음', '~펼침', '~임', '~함', '~됨'등으로 끝나는 명사형 종결어미를 사용하여 문장을 끝맺어야 함`
   }, {
     role: "assistant", content: "진로맞춤형 독서 활동을 통해 기계 공학자와 로봇 공학자에 대한 흥미와 가능성을 발견하고, 관련 분야에 대한 구체적이고 깊은 이해를 높임."
       + " 독서 모임에서 보고서 작성과 자료 정리를 맡아 팀원들이 제공한 자료를 주제별로 정리하고 취합해 탐구 보고서를 작성함. 토론 내용을 정리하고 관련 문서를 작성하며,"
@@ -451,13 +486,13 @@ const gptPersonalOnReportMsg = [
     독서를 통해 진로에 대한 흥미와 가능성을 발견하고, 해당 분야에 대한 깊은 이해를 높임. 진로 탐색을 통해 자신의 강점과 흥미를 발견하고,
     독서를 통해 쌓은 지식을 현실적인 진로 선택에 활용함. 함께한 동료들과의 토의를 통해 다양한 진로에 대한 시야를 넓히고, 진로 결정에 도움을 주는 활동을 진행함. 
 
-    아래는 위의 활동을 한 학생이 쓴 진로 보고서야.
+    아래는 위의 활동을 한 학생이 쓴 활동 보고서야.
 
-    활동 보고서: 독서활동에 자료조사 및 보고서 작성 역할을 맡음. "바이오닉 맨"을 읽으며 파일럿으로서 필요한 정신력과 위기 극복의 중요성을 깨닫고, 
+    학생 활동 보고서: 독서활동에 자료조사 및 보고서 작성 역할을 맡음. "바이오닉 맨"을 읽으며 파일럿으로서 필요한 정신력과 위기 극복의 중요성을 깨닫고, 
     파일럿의 훈련 과정에서 신체적, 정신적 성장의 필요성을 느꼈고,  또한 "이상한 미래 연구소"는 파일럿의 미래에 영향을 미칠 첨단 기술과 트렌드를 소개해 주면서,
     변화에 맞춰 준비하고 새로운 기술을 적극적으로 활용하는 파일럿이 되어야겠다는 생각을 했다.
 
-    위 활동과 보고서를 적절히 섞어서 생기부 문구를 만들어줘, 조건은 '학생'이라는 주어가 없어야 하고 '~음', '~펼침', '~임', '~함', '~됨'등으로 끝나는 문장으로 작성해줘`
+    위 활동과 보고서를 적절히 섞어서 생기부 문구를 만들어줘, 조건은 '학생'이라는 주어가 없어야 하고 문장을 끝낼 때, 반드시 '~음', '~펼침', '~임', '~함', '~됨'등으로 끝나는 명사형 종결어미를 사용하여 문장을 끝맺어야 함`
   }, {
     role: "assistant", content: "독서활동에서 자료 조사와 보고서 작성 역할을 맡아 팀의 완성도를 높이는 데 기여함."
       + " '바이오닉 맨'을 통해 파일럿으로서 필요한 정신력과 위기 극복의 중요성을 깨닫고, 훈련 과정에서 신체적·정신적 성장이 필수적임을 인식함."
@@ -471,16 +506,16 @@ const gptPersonalOnReportMsg = [
     독서를 통해 진로에 대한 흥미와 가능성을 발견하고, 해당 분야에 대한 깊은 이해를 높임. 진로 탐색을 통해 자신의 강점과 흥미를 발견하고,
     독서를 통해 쌓은 지식을 현실적인 진로 선택에 활용함. 함께한 동료들과의 토의를 통해 다양한 진로에 대한 시야를 넓히고, 진로 결정에 도움을 주는 활동을 진행함. 
 
-    아래는 위의 활동을 한 학생이 쓴 진로 보고서야.
+    아래는 위의 활동을 한 학생이 쓴 활동 보고서야.
 
-    활동 보고서: '다시, 수학이 필요한 순간'을 읽고 책에서 소개된 '불완전성 원리'에 흥미가 생겨 추가로 조사해보았다. 이를 바탕으로 '평범한 집합은 평범한가'라는 주제로
+    학생 활동 보고서: '다시, 수학이 필요한 순간'을 읽고 책에서 소개된 '불완전성 원리'에 흥미가 생겨 추가로 조사해보았다. 이를 바탕으로 '평범한 집합은 평범한가'라는 주제로
     스스로 탐구를 진행하면서, 평범함이라는 성질을 어떻게 정의할 수 있는지, 그리고 어떤 정리가 확실하며 또 어떤 정리는 확실하지 않은지에 대해 깊이 생각해보았다.
     이 과정을 통해 얻은 내용을 팀원들에게 발표하며 함께 토론했다. 또한, 심화 탐구 주제 중 하나였던 “무한 급수” 부분에서는 무한 급수의 수렴성을 판정할 수 있는 비교 판정법과
     비율 판정법에 대해 주도적으로 조사하였고, 이 판정법들이 실제로 어떻게 활용되는지와 그 증명 과정에 대해 팀원들에게 설명했다. 그리고 '떨림과 울림'에서 읽은 
     러더퍼드의 알파입자 산란 실험이 흥미로워 더 깊이 조사해 팀원들에게 소개했다. 팀원들과 함께한 자유 낙하 실험에서도 주도적으로 실험에 참여해 실험을 진행했고,
     실험 결과를 트래커 프로그램으로 분석하는 작업에도 적극적으로 참여하며 더 많은 것을 배우는 시간을 가졌다.
 
-    위 활동과 보고서를 적절히 섞어서 생기부 문구를 만들어줘, 조건은 '학생'이라는 주어가 없어야 하고 '~음', '~펼침', '~임', '~함', '~됨'등으로 끝나는 문장으로 작성해줘`
+    위 활동과 보고서를 적절히 섞어서 생기부 문구를 만들어줘, 조건은 '학생'이라는 주어가 없어야 하고 문장을 끝낼 때, 반드시 '~음', '~펼침', '~임', '~함', '~됨'등으로 끝나는 명사형 종결어미를 사용하여 문장을 끝맺어야 함`
   }, {
     role: "assistant", content: "독서 활동에서 '다시, 수학이 필요한 순간'을 읽고 흥미를 느낀 '불완전성 원리'를 바탕으로 '평범한 집합은 평범한가'라는 주제로"
       + " 스스로 탐구를 진행하며 평범함의 정의와 확실성의 개념에 대해 깊이 탐구함. 무한 급수의 수렴성을 판정하는 비교 판정법과 비율 판정법을 조사해 팀원들에게"
@@ -492,15 +527,15 @@ const gptPersonalOnReportMsg = [
     role: "user",
     content: `활동 문구: 모의 영어 면접 활동에서 생명공학과에 대한 열정을 영어로 표현하여 면접을 진행함. 직무 관련 질문에 대해 자신만의 시각과 경험을 바탕으로 대답함.
 
-    아래는 위의 활동을 한 학생이 쓴 진로 보고서야.
+    아래는 위의 활동을 한 학생이 쓴 활동 보고서야.
 
-    활동 보고서: 환경 과학, 이에 의한 사회적 문제, 생명 과학, 인공지능(딥러닝)을 집중적으로 탐구했다. 특히나 생명과학과 인공지능에 대해서 깊게 탐구해 보았다. 
+    학생 활동 보고서: 환경 과학, 이에 의한 사회적 문제, 생명 과학, 인공지능(딥러닝)을 집중적으로 탐구했다. 특히나 생명과학과 인공지능에 대해서 깊게 탐구해 보았다. 
     아직 확실한 꿈은 정해지지 않았지만 우리의 신체를 탐구하는 쪽이나 인공지능을 개발하는 일을 하고 싶다. 혹은 두 분야의 융합을 통해 우리의 신체를 분석해주는 인공지능을
     개발하는 일이라면 더욱 좋을 것 같다. 이 분야에서 나는 생명 과학의 경우, 생명의 탐구와 발전만을 고려하는 것이 아니라, 그 발전이 불러일으킬 영향, 탐구하는 과정에서 발생하는 역효과나
     부작용 등을 모두 고려해야 한다는 점을 잊지 않아야겠다고 생각했고, 인공지능의 경우, 초등학교와 중학교 때 서울교대 영재원과 서울대 영재원에서 탐구해온 인공지능과 딥러닝에 대해
     다시 한번 복습할 수 있는 계기가 되어 좋았다.
 
-    위 활동과 보고서를 적절히 섞어서 생기부 문구를 만들어줘, 조건은 '학생'이라는 주어가 없어야 하고 '~음', '~펼침', '~임', '~함', '~됨'등으로 끝나는 문장으로 작성해줘`
+    위 활동과 보고서를 적절히 섞어서 생기부 문구를 만들어줘, 조건은 '학생'이라는 주어가 없어야 하고 문장을 끝낼 때, 반드시 '~음', '~펼침', '~임', '~함', '~됨'등으로 끝나는 명사형 종결어미를 사용하여 문장을 끝맺어야 함`
   }, {
     role: "assistant", content: "모의 영어 면접 활동에서 자신의 진로에 대한 열정을 영어로 표현하고, 직무 관련 질문에 대해 생명과학과 인공지능 분야에 대한 자신만의 시각과 경험을 바탕으로 대답함."
       + " 환경 과학, 사회적 문제, 생명 과학, 인공지능(딥러닝)에 집중적으로 탐구하며, 생명과학과 인공지능 분야에서의 꿈을 고민하고 있다고 영어로 답함."
@@ -513,15 +548,15 @@ const gptPersonalOnReportMsg = [
     role: "user",
     content: `활동 문구: 모의 영어 면접 활동에서 생명공학과에 대한 열정을 영어로 표현하여 면접을 진행함. 직무 관련 질문에 대해 자신만의 시각과 경험을 바탕으로 대답함.
   
-      아래는 위의 활동을 한 학생이 쓴 진로 보고서야.
+      아래는 위의 활동을 한 학생이 쓴 활동 보고서야.
   
-      활동 보고서: 저는 의사가 되어 생명을 살리고, 좀 더 이 사회를 따뜻하게 만들고 싶다는 희망을 가지고 있습니다. 봉사와 사랑을 실천하는 의사들의 삶이 궁금하였고 
+      학생 활동 보고서: 저는 의사가 되어 생명을 살리고, 좀 더 이 사회를 따뜻하게 만들고 싶다는 희망을 가지고 있습니다. 봉사와 사랑을 실천하는 의사들의 삶이 궁금하였고 
       그 자세를 본받고 싶어서 이 주제를 선택하게 되었습니다. 이들은 단순히 안정된 직업과 명예 때문에 의사가 된 것이 아니라 ‘다른 사람을 도우려는 마음’에서 어려운 공부를 하였고,
       적은 대가를 받으면서도 그 일을 수행하고 있다는 것을 알게 되었습니다. 생명과 밀접한 관련이 있는 의사라는 직업은 높은 도덕 의식과 봉사 정신이 필요하다는 것을 
       이번 면접을 준비하며 깨닫게 되었습니다. 앞으로 의사가 된다면 열심히 공부하여 전문적 지식과 기술을 갖추고 다른 사람의 고통을 덜어주며 생명을 살리는 일을 하고 싶습니다.
       다른 사람들에게 도움이 되는 것에 보람을 느끼고 꾸준히 봉사와 사랑을 실천하는 가슴 따뜻한 의사가 되고 싶습니다.
 
-      위 활동과 보고서를 적절히 섞어서 생기부 문구를 만들어줘, 조건은 '학생'이라는 주어가 없어야 하고 '~음', '~펼침', '~임', '~함', '~됨'등으로 끝나는 문장으로 작성해줘`
+      위 활동과 보고서를 적절히 섞어서 생기부 문구를 만들어줘, 조건은 '학생'이라는 주어가 없어야 하고 문장을 끝낼 때, 반드시 '~음', '~펼침', '~임', '~함', '~됨'등으로 끝나는 명사형 종결어미를 사용하여 문장을 끝맺어야 함`
   }, {
     role: "assistant", content: "모의 영어 면접 활동에서 의대로 진학하고자 하는 면접을 영어로 진행하여 의학에 대한 열정과 관련 역량을 자신의 경험과 비전을 바탕으로 표현함."
       + " 진학 동기와 비전, 업무 역량에 대한 질문에 책에서 얻은 통찰을 바탕으로 '봉사와 사랑을 실천하는 의사들의 삶이 궁금했고, 그들의 마음가짐을 본받고 싶어서 의학을 선택했습니다'"
@@ -691,4 +726,60 @@ const gptPerfRecordMsg = [
       + "삶과 죽음에 관한 연사들의 연설문을 읽고 외워 그 중 일부분을 영어로 발표함."
   }
 ]
-export { gptBehaviorMsg, gptSubjectDetailsMsg, gptHomeroomDetailMsg, gptPersonalOnReportMsg, gptPersonalOnTraitMsg, gptExtraRecordMsg, gptPerfRecordMsg }
+const gptTranslateMsg = [
+  { role: "system", content: `당신은 전문 번역가입니다. 주어진 문장을 자연스럽게 한국말로 번역해주세요. 학생들이 쓴 글이라 문법적 오류가 많을 수 있습니다. 말이 되지 않는 부분은 맥락에 맞게 적절히 해석해주세요.` },
+  //예시1
+  {
+    role: "user", content: `
+    I think 'Lamb' symbolizes Mary Maloney. She was a very kind, gentle and beautiful house wife. But when she kills her husband, she tuns into a slaughterman. It was very dramatic. The situational ironies is how the detectives 
+    eat the lamb's leg which is the evidence for Mary Maloney's murder.,
+    위의 문장을 자연스러운 한국말로 번역해주세요.
+  `
+  }, {
+    role: "assistant", content: `나는 '양'이 메리 말로니를 상징한다고 생각해요. 그녀는 매우 친절하고 온순하며 아름다운 주부였어요. 
+    그러나 남편을 죽인 후에는 도살자로 변모했죠. 매우 드라마틱한 일이었어요. 상황적 역설은 형사들이 메리 말로니의 살인을 증거하는 양다리를 먹는 것입니다.`
+  },
+  //예시2
+  {
+    role: "user", content: `
+    I think the lamb symbolizes Mary. In the bible, usually the lamb symbolizes pure, innocent child of God. But in the novel, Mary becomes a murderer. So the lamb is going to the soughter, but the lamb it's self Is the slaughter.
+    There are I Tronic situations in this novel. Frist, Mary killing her own hushand. In the opening of the novel, Mary waits for her husband with excitement. But later she kills he beloved hushand with her own hands.
+    Second, she uses a food a leg of the lamb to kill him. And she even destoryed the evidence with the detectives. I really got goosebumps when Mary made the detectives eat the lamb and giggled.,
+    위의 문장을 자연스러운 한국말로 번역해주세요.
+  `
+  }, {
+    role: "assistant", content: `양은 메리를 상징한다고 생각해요. 성경에서 양은 보통 하나님의 순수하고 결백한 아이를 상징합니다. 
+    그러나 소설에서 메리는 살인자가 됩니다. 그래서 양은 도살당하러 가는 것이지만, 양 자체가 도살당하는 존재가 됩니다. 이 소설에는 역설적인 상황들이 있어요. 
+    먼저, 메리가 자신의 남편을 죽입니다. 소설의 시작에서 메리는 남편을 기대하며 기다리지만 나중에는 그녀가 사랑하는 남편을 자신의 손으로 죽입니다. 
+    둘째, 그녀는 양다리를 사용해 그를 죽입니다. 그리고 그녀는 형사들이 증거를 파괴하도록 합니다. 메리가 형사들에게 양을 먹이고 웃는 장면에서 정말 오싹한 느낌이었어요.`
+  },
+  //예시2
+  {
+    role: "user", content: `
+    I think the lamb symbolizes Mary. In the bible, usually the lamb symbolizes pure, innocent child of God. But in the novel, Mary becomes a murderer. So the lamb is going to the soughter, but the lamb it's self Is the slaughter.
+    There are I Tronic situations in this novel. Frist, Mary killing her own hushand. In the opening of the novel, Mary waits for her husband with excitement. But later she kills he beloved hushand with her own hands.
+    Second, she uses a food a leg of the lamb to kill him. And she even destoryed the evidence with the detectives. I really got goosebumps when Mary made the detectives eat the lamb and giggled.,
+    위의 문장을 자연스러운 한국말로 번역해주세요.
+  `
+  }, {
+    role: "assistant", content: `양은 메리를 상징한다고 생각해요. 성경에서 양은 보통 하나님의 순수하고 결백한 아이를 상징합니다. 
+    그러나 소설에서 메리는 살인자가 됩니다. 그래서 양은 도살당하러 가는 것이지만, 양 자체가 도살당하는 존재가 됩니다. 이 소설에는 역설적인 상황들이 있어요. 
+    먼저, 메리가 자신의 남편을 죽입니다. 소설의 시작에서 메리는 남편을 기대하며 기다리지만 나중에는 그녀가 사랑하는 남편을 자신의 손으로 죽입니다. 
+    둘째, 그녀는 양다리를 사용해 그를 죽입니다. 그리고 그녀는 형사들이 증거를 파괴하도록 합니다. 메리가 형사들에게 양을 먹이고 웃는 장면에서 정말 오싹한 느낌이었어요.`
+  },
+  //예시3
+  {
+    role: "user", content: `
+    Usually. The lamb means pure and innocent. But the camb symbolizes both sucrifice and violence in the story. Lamb symbolizes Mary's purity in the first part but lamb changes into Sanguinary Cruelty as a weapon for murder.
+    First situational irony on the story can be find from Mary. The loving wife who is guessed to be victim transformed into cruel detective's consumption of evidence, the leg of lamb. 
+    The detective murderer killing hushand second irony is Investigate the crime ate the evidence of murder.,
+    위의 문장을 자연스러운 한국말로 번역해주세요.
+  `
+  }, {
+    role: "assistant", content: `보통 '양'은 순수하고 결백함을 의미합니다. 그러나 이 이야기에서는 양이 희생과 폭력을 상징합니다.
+    양은 처음에는 메리의 순수함을 상징하지만, 살인 무기로 변모한 피의 잔혹함이 되었습니다. 이 이야기에서 첫 번째 상황적 역설은 메리에서 찾을 수 있습니다.
+    피해자로 추정되는 사랑스러운 아내가 잔혹한 형사들의 증거물인 양다리를 먹이로 변해버렸습니다. 
+    두 번째 역설은 범죄를 조사하던 형사가 살인의 증거물을 먹는 것입니다.`
+  }
+]
+export { gptBehaviorMsg, gptSubjectDetailsMsg, gptHomeroomDetailMsg, gptPersonalOnReportMsg, gptPersonalOnTraitMsg, gptExtraRecordMsg, gptPerfRecordMsg, gptTranslateMsg }
